Update Currency
CB - Updated delegate from being dynamic to not because I didn't to use it in blueprints. Updated the currency amount to an int instead of a float

+++ Source/TowerDefense/HUDPlayerCurrency.h
@@ -17,5 +17,5 @@ public:
 	UTextBlock* CurrencyText;
 
 	UFUNCTION(BlueprintCallable)
-	void SetTextToPlayerCurrency(float amount);
+	void SetTextToPlayerCurrency(int amount);
 };
